Course-Member structure
Added sister group field to models.py

Courseaddview now automatically assigns instructor as its creator. Creates sister group when a course is created and assigns the instructor to that group. In order to be 'part' of a course, one must be part of its sister group

Created view for studentPage to handle course logic

Student page now displays all courses a student is enrolled in (part of sister group), as well as a distinction on whether or not that user is an instructor

Configured bootstrap for course_add.html. Changed filepath for this template as well, now handled under project

Slight adjustments to home page (auth logic)